// 6. semestr/[IBT] Bakalářská práce/Implementace/frontend/src/app/models/rendering/selected-person.ts
/*******************************************************************/
/*  Project: Bachelor's thesis – Genealogy Database System         */
/*           Faculty of Information Technology                     */
/*           Brno University of Technology                         */
/*                                                                 */
/*  File:    selected-person.ts                                    */
/*  Date:    2019 – 2020                                           */
/*  Author:  Martin Chládek <xchlad16@stud.fit.vutbr.cz>           */
/*******************************************************************/

import { Person } from '../person';

export interface SelectedPerson {

    id: number;
    name: string;
    data: Person;
}